clang-tidy

+++ esphome/components/i80/i80_component.cpp
@@ -7,7 +7,7 @@ void I80Component::dump_config() {
   ESP_LOGCONFIG(TAG, "I80 bus:");
   LOG_PIN("  WR Pin: ", this->wr_pin_)
   LOG_PIN("  DC Pin: ", this->dc_pin_)
-  for (size_t i = 0; i != this->data_pins_.size(); i++) {
+  for (unsigned i = 0; i != this->data_pins_.size(); i++) {
     ESP_LOGCONFIG(TAG, "  Data pin %u: GPIO%d", i, this->data_pins_[i]);
   }
 }
